+++ src/language/languages.ts
@@ -1,6 +1,7 @@
 import { csharpLanguage } from './csharpLanguage.js'
 import { javaLanguage } from './javaLanguage.js'
 import { phpLanguage } from './phpLanguage.js'
+import { pythonLanguage } from './pythonLanguage.js'
 import { rubyLanguage } from './rubyLanguage.js'
 import { Language, LanguageName } from './types.js'
 import { typescriptLanguage } from './typescriptLanguage.js'
@@ -11,6 +12,7 @@ const languageByName: Record<LanguageName, Language> = {
   c_sharp: csharpLanguage,
   php: phpLanguage,
   ruby: rubyLanguage,
+  python: pythonLanguage,
 }
 
 export function getLanguage(languageName: LanguageName): Language {